pyproject.toml
[project]
name = "Mutator"
version = "0.1.0"
description = "A LLM powered mutation testing tool"
readme = "README.md"
# TODO: License
requires-python = ">=3.10"
dependencies = [
	"tree-sitter>=0.22.0",
	"tree-sitter-python>=0.21.0",
	"pytest>=8.2.0",
	"accelerate>=0.30.0",
	"sentencepiece>=0.2.0",
	"torch>=2.3.0",
	"transformers>=4.40.1",
	"textual>=0.59.0",
	"GitPython>=3.1.43",
	"click>=8.1.7",
	"autopep8>=2.2.0",
	"datasets>=2.19.1",
	"peft>=0.11.1",
	"pandas>=2.2.2",
	"matplotlib>=3.9.0",
	"setuptools>=70.0.0",
	# "evaluate>=0.4.2",
	# "numpy>=1.26.4",
	# "scikit-learn>=1.5.0"
]

[project.scripts]
mutator = "mutator:cli_main"
mutator-runner = "mutator_runner:cli_main"

[tool.pyright]
pythonVersion = "3.12"
include = ["src/mutator", "tests"]
typeCheckingMode = "basic"

[tool.ruff]
src = ["src"]
fix = true
show-fixes = true
output-format = "full"

[tool.ruff.lint]
select = [ "B", "E", "F", "I", "UP", "W" ]
